ApiClient

ApiClient Class

Format

An R6Class generator object

Details

ApiClient Class

Generic API client for OpenAPI client library builds. OpenAPI generic API client. This client handles the client- server communication, and is invariant across implementations. Specifics of the methods and models for each application are generated from the OpenAPI Generator templates.

NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech). Ref: https://openapi-generator.tech Do not edit the class manually.

Public fields

basePath

Base url

userAgent

Default user agent

username

Username for HTTP basic authentication

password

Password for HTTP basic authentication

timeout

Default timeout in seconds

Methods

Public methods

ApiClient$new()

ApiClient$CallApi()

ApiClient$deserialize()

ApiClient$deserializeObj()

ApiClient$clone()

Method new()

Usage

ApiClient$new(
  basePath = NULL,
  userAgent = NULL,
  defaultHeaders = NULL,
  username = NULL,
  password = NULL,
  apiKeys = NULL,
  accessToken = NULL,
  timeout = NULL
)

Method CallApi()

Usage

ApiClient$CallApi(url, method, queryParams, headerParams, body, ...)

Method deserialize()

Usage

ApiClient$deserialize(resp, returnType, pkgEnv)

Method deserializeObj()

Usage

ApiClient$deserializeObj(obj, returnType, pkgEnv)

Method clone()

The objects of this class are cloneable with this method.

Usage

ApiClient$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.