Metadata

Metadata Class

Format

An R6Class generator object

Public fields

prefix

character

mzTab-version

character

mzTab-ID

character

title

character [optional]

description

character [optional]

sample_processing

list( SampleProcessing ) [optional]

instrument

list( Instrument ) [optional]

software

list( Software )

publication

list( Publication ) [optional]

contact

list( Contact ) [optional]

uri

list( Uri ) [optional]

external_study_uri

list( Uri ) [optional]

quantification_method

Parameter

study_variable

list( StudyVariable )

ms_run

list( MsRun )

assay

list( Assay )

sample

list( Sample ) [optional]

custom

list( Parameter ) [optional]

cv

list( CV )

database

list( Database )

derivatization_agent

list( Parameter ) [optional]

small_molecule-quantification_unit

Parameter

small_molecule_feature-quantification_unit

Parameter

small_molecule-identification_reliability

Parameter [optional]

id_confidence_measure

list( Parameter )

colunit-small_molecule

list( ColumnParameterMapping ) [optional]

colunit-small_molecule_feature

list( ColumnParameterMapping ) [optional]

colunit-small_molecule_evidence

list( ColumnParameterMapping ) [optional]

Methods

Public methods

Metadata$new()

Metadata$toJSON()

Metadata$fromJSON()

Metadata$toJSONString()

Metadata$fromJSONString()

Metadata$toDataFrame()

Metadata$fromDataFrame()

Metadata$clone()

Method new()

Create Metadata

Usage

Metadata$new(
  prefix,
  `mzTab-version`,
  `mzTab-ID`,
  software,
  quantification_method,
  study_variable,
  ms_run,
  assay,
  cv,
  database,
  `small_molecule-quantification_unit`,
  `small_molecule_feature-quantification_unit`,
  id_confidence_measure,
  title = NULL,
  description = NULL,
  sample_processing = NULL,
  instrument = NULL,
  publication = NULL,
  contact = NULL,
  uri = NULL,
  external_study_uri = NULL,
  sample = NULL,
  custom = NULL,
  derivatization_agent = NULL,
  `small_molecule-identification_reliability` = NULL,
  `colunit-small_molecule` = NULL,
  `colunit-small_molecule_feature` = NULL,
  `colunit-small_molecule_evidence` = NULL,
  ...
)

Arguments

prefix

'MTD'.

mzTab-version

The version e.g. '2.0.0-M'.

mzTab-ID

MzTab file id.

software

Software used to generate the results and the file list( Software ).

quantification_method

Quantification method Parameter.

study_variable

Study variables list( StudyVariable ).

ms_run

MsRuns of this study list( MsRun ).

assay

Assays linking Samples and MsRuns list( Assay ).

cv

Controlled vocabularies list( CV ).

database

Databases used for identification or matching list( Database ).

small_molecule-quantification_unit

Quantification unit for small molecule summary abundances Parameter.

small_molecule_feature-quantification_unit

Quantification unit for small molecule feature abundances Parameter.

id_confidence_measure

Identification confidence rule set ( Parameter ).

title

The title of the MzTab file.

description

The description.

sample_processing

List of sample processing steps list( SampleProcessing ).

instrument

Instruments used to acquire MsRuns list( Instrument ).

publication

Publications linked to this data list( Publication ).

contact

Contacts / creators of this file list( Contact ).

uri

External URIs relevant to this file list( Uri ).

external_study_uri

Where to find this study externall list( Uri ).

sample

Samples of this study list( Sample ).

custom

Custom definitions for this study list( Parameter ).

derivatization_agent

Derivatization agents used list( Parameter ).

small_molecule-identification_reliability

Identification reliability rule set Parameter.

colunit-small_molecule

Additional SmallMoleculeSummary column units list( ColumnParameterMapping ).

colunit-small_molecule_feature

Additional SmallMoleculeFeature column units list( ColumnParameterMapping ).

colunit-small_molecule_evidence

Additional SmallMoleculeEvidence column units list( ColumnParameterMapping ).

...

local optional variable arguments

Method toJSON()

Serialize to list object suitable for jsonlite

Usage

Metadata$toJSON()

Method fromJSON()

Deserialize from jsonlite list object

Usage

Metadata$fromJSON(MetadataJson)

Arguments

MetadataJson

list object.

Method toJSONString()

Serialize to JSON string.

Usage

Metadata$toJSONString()

Method fromJSONString()

Deserialize from JSON string

Usage

Metadata$fromJSONString(MetadataJson)

Arguments

MetadataJson

JSON string

Method toDataFrame()

Serialize to data frame

Usage

Metadata$toDataFrame()

Method fromDataFrame()

Deserialize from metadata data frame

Usage

Metadata$fromDataFrame(MetadataDataFrame)

Arguments

MetadataDataFrame

Metadata data frame

Method clone()

The objects of this class are cloneable with this method.

Usage

Metadata$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.