SmallMoleculeSummary Class

The small molecule summary section is table-based, represented summarized quantitative information across assays and study variables, grouped by identification in rows. The small molecule section MUST always come after the metadata section. All table columns MUST be Tab separated. There MUST NOT be any empty cells; missing values MUST be reported using “null” for columns where Is Nullable &#x3D; “True”. Each row of the small molecule section is intended to report one final result to be communicated in terms of a molecule that has been quantified. In many cases, this may be the molecule of biological interest, although in some cases, the final result could be a derivatized form as appropriate – although it is desirable for the database identifier(s) to reference to the biological (non-derivatized) form. In general, different adduct forms would generally be reported in the Small Molecule Feature section. The order of columns MUST follow the order specified below. All columns are MANDATORY except for “opt_” columns.

SmallMoleculeSummary

Format

An object of class R6ClassGenerator of length 24.

Fields

prefix

header_prefix

sml_id

smf_id_refs

database_identifier

chemical_formula

smiles

inchi

chemical_name

uri

theoretical_neutral_mass

adduct_ions

reliability

best_id_confidence_measure

best_id_confidence_value

abundance_assay

abundance_study_variable

abundance_variation_study_variable

opt

comment