ValidatePlain operations

openapi.ValidatePlain

Format

An R6Class generator object

Methods

ValidatePlainMzTabFile Validates an mzTab file in plain text representation / tab-separated format and reports syntactic, structural, and semantic errors. @param mztabfile character @param level Enum < [info, warn, error] > @param max.errors integer @param semantic.validation character @returnType list( ValidationMessage )
status code : 200 | Validation Okay return type : array[ValidationMessage] response headers :

status code : 415 | Unsupported content type response headers :

status code : 422 | Invalid input return type : array[ValidationMessage] response headers :

status code : 0 | Unexpected error return type : Error response headers :

Public fields

apiClient

Handles the client-server communication.

Methods

Public methods

ValidatePlainApi$new()

ValidatePlainApi$ValidatePlainMzTabFile()

ValidatePlainApi$ValidatePlainMzTabFileWithHttpInfo()

ValidatePlainApi$clone()

Method new()

Usage

ValidatePlainApi$new(apiClient)

Method ValidatePlainMzTabFile()

Usage

ValidatePlainApi$ValidatePlainMzTabFile(
  mztabfile,
  level = "info",
  max.errors = 100,
  semantic.validation = FALSE,
  ...
)

Method ValidatePlainMzTabFileWithHttpInfo()

Usage

ValidatePlainApi$ValidatePlainMzTabFileWithHttpInfo(
  mztabfile,
  level = "info",
  max.errors = 100,
  semantic.validation = FALSE,
  ...
)

Method clone()

The objects of this class are cloneable with this method.

Usage

ValidatePlainApi$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (FALSE) {
####################  ValidatePlainMzTabFile  ####################

library(openapi)
var.mztabfile <- 'mztabfile_example' # character | mzTab file that should be validated.
var.level <- 'info' # character | The level of errors that should be reported, one of ERROR, WARN, INFO.
var.max.errors <- 100 # integer | The maximum number of errors to return.
var.semantic.validation <- FALSE # character | Whether a semantic validation against the default rule set should be performed.

api.instance <- ValidatePlainApi$new()

result <- api.instance$ValidatePlainMzTabFile(var.mztabfile, level=var.level, max.errors=var.max.errors, semantic.validation=var.semantic.validation)

}