Converts the provided Metadata object to a ragged data frame.

Converts the provided Metadata object to a ragged data frame.

as.data.frame(metadata)

Arguments

metadata
the R6 Metadata object to write.