MsRun Class

Specification of ms_run. location: Location of the external data file e.g. raw files on which analysis has been performed. If the actual location of the MS run is unknown, a “null” MUST be used as a place holder value, since the [1-n] cardinality is referenced elsewhere. If pre-fractionation has been performed, then [1-n] ms_runs SHOULD be created per assay. instrument_ref: If different instruments are used in different runs, instrument_ref can be used to link a specific instrument to a specific run. format: Parameter specifying the data format of the external MS data file. If ms_run[1-n]-format is present, ms_run[1-n]-id_format SHOULD also be present, following the parameters specified in Table 1. id_format: Parameter specifying the id format used in the external data file. If ms_run[1-n]-id_format is present, ms_run[1-n]-format SHOULD also be present. fragmentation_method: The type(s) of fragmentation used in a given ms run. scan_polarity: The polarity mode of a given run. Usually only one value SHOULD be given here except for the case of mixed polarity runs. hash: Hash value of the corresponding external MS data file defined in ms_run[1-n]-location. If ms_run[1-n]-hash is present, ms_run[1-n]-hash_method SHOULD also be present. hash_method: A parameter specifying the hash methods used to generate the String in ms_run[1-n]-hash. Specifics of the hash method used MAY follow the definitions of the mzML format. If ms_run[1-n]-hash is present, ms_run[1-n]-hash_method SHOULD also be present.

Public fields

id

elementType

name

location

instrument_ref

format

id_format

fragmentation_method

scan_polarity

hash

hash_method

mzTab-M specification example(s)

COM    location can be a local or remote URI
MTD    ms_run[1]-location    file:///C:/path/to/my/file.mzML
MTD    ms_run[1]-instrument_ref    instrument[1]
MTD    ms_run[1]-format    [MS, MS:1000584, mzML file, ]
MTD    ms_run[1]-id_format    [MS, MS:1000530, mzML unique identifier, ]
MTD    ms_run[1]-fragmentation_method[1]    [MS, MS:1000133, CID, ]
COM    for mixed polarity scan scenarios
MTD    ms_run[1]-scan_polarity[1]    [MS, MS:1000130, positive scan, ]
MTD    ms_run[1]-scan_polarity[2]    [MS, MS:1000129, negative scan, ]
MTD    ms_run[1]-hash_method    [MS, MS:1000569, SHA-1, ]
MTD    ms_run[1]-hash    de9f2c7fd25e1b3afad3e85a0bd17d9b100db4b3

Methods

Public methods

MsRun$new()

MsRun$toJSON()

MsRun$fromJSON()

MsRun$toJSONString()

MsRun$fromJSONString()

MsRun$clone()

Method new()

Usage

MsRun$new(
  id,
  elementType,
  name,
  location,
  instrument_ref,
  format,
  id_format,
  fragmentation_method,
  scan_polarity,
  hash,
  hash_method
)

Method toJSON()

Usage

MsRun$toJSON()

Method fromJSON()

Usage

MsRun$fromJSON(MsRunJson)

Method toJSONString()

Usage

MsRun$toJSONString()

Method fromJSONString()

Usage

MsRun$fromJSONString(MsRunJson)

Method clone()

The objects of this class are cloneable with this method.

Usage

MsRun$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.