Validate operations

rmzTabM.Validate

Format

An R6Class generator object

Methods

ValidateMzTabFile Validates an mzTab file in XML or JSON representation and reports syntactic, structural, and semantic errors. @param mztabfile MzTab @param level Enum < [info, warn, error] > @param max.errors integer @param semantic.validation character @returnType list( ValidationMessage )
status code : 200 | Validation Okay return type : array[ValidationMessage] response headers :

status code : 415 | Unsupported content type response headers :

status code : 422 | Invalid input return type : array[ValidationMessage] response headers :

status code : 0 | Unexpected error return type : Error response headers :

Public fields

apiClient

Handles the client-server communication.

Methods

Public methods

ValidateApi$new()

ValidateApi$ValidateMzTabFile()

ValidateApi$ValidateMzTabFileWithHttpInfo()

ValidateApi$clone()

Method new()

Create a new ValidateApi object.

Usage

ValidateApi$new(apiClient)

Arguments

apiClient

An initialized ApiClient.

Method ValidateMzTabFile()

Validate an MzTab object.

Usage

ValidateApi$ValidateMzTabFile(
  mztabfile,
  level = "info",
  max.errors = 100,
  semantic.validation = FALSE,
  ...
)

Arguments

mztabfile

The MzTab object to validate.

level

The validation level.

max.errors

The maximum number of errors.

semantic.validation

Whether semantic validation should also be performed.

...

local optional variable arguments

Method ValidateMzTabFileWithHttpInfo()

Validate an MzTab file but include HTTP response info.

Usage

ValidateApi$ValidateMzTabFileWithHttpInfo(
  mztabfile,
  level = "info",
  max.errors = 100,
  semantic.validation = FALSE,
  ...
)

Arguments

mztabfile

The MzTab object to validate.

level

The validation level.

max.errors

The maximum number of errors.

semantic.validation

Whether semantic validation should also be performed.

...

local optional variable arguments

Method clone()

The objects of this class are cloneable with this method.

Usage

ValidateApi$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (FALSE) {
####################  ValidateMzTabFile  ####################

library(rmzTabM)
# MzTab | mzTab file that should be validated.
var.mztabfile <- MzTab$new()
# character | The level of errors that should be reported, one of ERROR, WARN, INFO.
var.level <- 'info' 
# integer | The maximum number of errors to return.
var.max.errors <- 100 
# character | Whether a semantic validation against the default rule set should be performed.
var.semantic.validation <- FALSE 

api.instance <- ValidateApi$new()

result <- api.instance$ValidateMzTabFile(var.mztabfile, 
  level=var.level, max.errors=var.max.errors, 
  semantic.validation=var.semantic.validation
)

}